+++ CHANGES.md
@@ -4,9 +4,11 @@
 This change actually severely breaks the API, but since no one else is using this crate yet
 I figure it's safe to deviate a little from semantic versioning best practices.
 * made all configuration methods chainable
-* made addrow utf8-safe
-* change macerate so it returns rows of lines of cells, the better to keep the mapping between
+* made `addrow` utf8-safe
+* change `macerate` so it returns rows of lines of cells, the better to keep the mapping between
 rows added and pieces returned
+* made `lay_out` private and added `reset` method to provide equivalent functionality
+* gave `tabulate` and `macerate` still more flexible signatures
 ## 1.0.0
 * simpler, more concise API
 * vertical alignment